web/requirements.txt
Flask
lxml
GoogleAppEngineCloudStorageClient
cachetools
